modify a bug about getting time and add proc file to debug some bugs for battery driver

+++ include/linux/power/jz4780-battery.h
@@ -37,7 +37,7 @@ enum {
 };
 
 struct private {
-	int timecount;
+	unsigned long timecount;
 	unsigned int voltage;
 };
